requirements-test.txt
pytest>=7.0.0
inline-snapshot>=0.12.0
Pillow>=9.0.0